Project trash (#100)
* Mode detector to thumbnail

* Open/close menu on longpress

* Anchor popup on top center

* Rotate arrow

* Rename triangle to arrow

* Doc

* Add arrow position

* Implement arrow on bottom option

* Update docs

* Remove arrow offset

* Handle horizontal start/end

* Handle arrow in the middle of horizontal side

* Start over

* Adjust for half pixel gap

* Handle horizontal positions

* Position arrow relative to contents

* Default to middle

* Finalize arrow position

* Extract width from popup

* Fix build

* Explicit top middle

* Menu top arrow position

* Align back drawing menu

* Adjust const

* Add todos

* Use popupwitharrow entry on thumbnail

* Anchor arrow

* Better var name

* Fix arrow anchor for middle side position

* Rename var

* Handle vertical sides

* Use popup with arrow entry for brush picker

* Optional popup color

* Extract labeled icon button

* Adjust arrow anchor on thumbnail

* Trash button in home bar

* Clip thumbnail corners

* Extract trash button

* Open/close trash

* Rename trash to bin

* Pass keys to project thumbnails

* Move project dir name knowledge to project class

* Move project creation knowledge to project class

* Rename id to creation epoch

* Grammar

* Extract dir name generator

* Change type

* Extract parsing creation epoch

* Add parsing of binned

* Extract binned postfix

* Dir name checker doesn't duplicate knowledge

* Move to bin method

* Project have to binned before deletion

* Add projects and binned projects getters

* Show binned projects in the bin

* Save project thumbnail with background

* Style bin thumbnails

* Destroy slidable button

* Restore slidable button

* Remove unused import

* Extract bin slide action buton

* Adjust splash color

* Handle destroying project

* Handle restoring project

* Write thumbnail even if frame is empty

* Show empty trash message

* Change restore icon

* Fade popup in/out

* Remove drawing menu extra padding

* Refactor frame menu with popup entry

* Remove unused color picker

* Bigger bin popup on larger screens

* Fix hitbox offset bug

* Fix frame resize bug

* Close popups on stroke start

* Simplify project deletion

* Rename var

* Remove unnecessary center

* Extract bin item widget

* Fix logo sharpness

+++ lib/gallery/data/gallery_model.dart
@@ -5,9 +5,13 @@ import 'package:path/path.dart' as p;
 
 class GalleryModel extends ChangeNotifier {
   Directory _directory;
-  List<Project> _projects;
+  List<Project> _projects = [];
 
-  int get numberOfProjects => _projects?.length;
+  List<Project> get projects =>
+      _projects.where((project) => !project.binned).toList();
+
+  List<Project> get binnedProjects =>
+      _projects.where((project) => project.binned).toList();
 
   Future<void> init(Directory directory) async {
     _directory = directory;
@@ -18,40 +22,59 @@ class GalleryModel extends ChangeNotifier {
   Future<List<Project>> _readProjects() async {
     final List<Project> projects = [];
     await for (final FileSystemEntity dir in _directory.list()) {
-      if (dir is Directory && _validProjectDirectory(dir)) {
+      if (dir is Directory && Project.validProjectDirectory(dir)) {
         projects.add(Project(dir));
       }
     }
     // Recent projects first.
-    projects.sort((p1, p2) => p2.id - p1.id);
+    projects.sort((p1, p2) => p2.creationEpoch - p1.creationEpoch);
     return projects;
   }
-
-  bool _validProjectDirectory(Directory directory) {
-    final String folderName = p.basename(directory.path);
-    return RegExp(r'^project_[0-9]+$').hasMatch(folderName);
-  }
-
-  Project getProject(int index) => _projects[index];
 
   Future<Project> addProject() async {
     if (_projects == null) throw Exception('Read projects first.');
 
-    final int id = DateTime.now().millisecondsSinceEpoch;
-    final String folderName = 'project_$id';
-    final Directory dir =
-        await Directory(p.join(_directory.path, folderName)).create();
-    final Project project = Project(dir);
-
+    final project = Project.createIn(_directory);
     _projects.insert(0, project);
 
     notifyListeners();
     return project;
   }
 
-  Future<void> deleteProject(int index) async {
-    final Project project = _projects.removeAt(index);
+  Future<void> moveProjectToBin(Project project) async {
+    if (!_projects.contains(project))
+      throw Exception('Project instance not found.');
+
+    final i = _projects.indexOf(project);
+    final newDirName = Project.directoryName(project.creationEpoch, true);
+    final newProjectPath = p.join(_directory.path, newDirName);
+
+    _projects[i] = Project(project.directory.renameSync(newProjectPath));
+    notifyListeners();
+  }
+
+  Future<void> deleteProject(Project project) async {
+    if (!_projects.contains(project))
+      throw Exception('Project instance not found.');
+    if (!project.binned)
+      throw Exception('Cannot delete project outside of bin.');
+
+    _projects.remove(project);
     await project.directory.delete(recursive: true);
+    notifyListeners();
+  }
+
+  Future<void> restoreProject(Project project) async {
+    if (!_projects.contains(project))
+      throw Exception('Project instance not found.');
+    if (!project.binned)
+      throw Exception('Cannot restore project outside of bin.');
+
+    final i = _projects.indexOf(project);
+    final newDirName = Project.directoryName(project.creationEpoch, false);
+    final newProjectPath = p.join(_directory.path, newDirName);
+
+    _projects[i] = Project(project.directory.renameSync(newProjectPath));
     notifyListeners();
   }
 }
